requirements.txt
numpy
pandas
scikit-learn
python-docx
typer
PyPDF2
pdfplumber
pytesseract
Pillow
networkx
matplotlib
requests
beautifulsoup4
transformers

# PDF Parsing
PyMuPDF==1.23.8
pdf2image==1.17.0
pytesseract==0.3.10
pydantic==2.5.3
python-dotenv==1.0.1

# LLM APIs
openai==1.7.2
google-generativeai==0.3.2
langchain==0.1.0

# Data Science & Machine Learning
pandas==2.1.4
numpy==1.26.3
scikit-learn==1.4.0
xgboost==2.0.3
shap==0.44.0

# Backend Output
fastapi==0.108.0
uvicorn==0.25.0
